+++ packages/tsurlfilter/CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.5.0] - 2025-11-20
+
+### Added
+
+- Support for regexp patterns in cosmetic rules.
+- Support for regexp in `domain` modifier [#105].
+
+### Changed
+
+- Updated [@adguard/agtree] to `v3.3.0`.
+
+[3.5.0]: https://github.com/AdguardTeam/tsurlfilter/releases/tag/tsurlfilter-v3.5.0
+[#105]: https://github.com/AdguardTeam/tsurlfilter/issues/105
+
 ## [3.4.8] - 2025-11-12
 
 ### Changed